Separate definition of event handler from Repair class
