+++ src/main/java/sorald/Repair.java
@@ -7,6 +7,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.EnumMap;
 import java.util.LinkedList;
 import java.util.List;
@@ -49,68 +50,16 @@ public class Repair {
     private final SoraldConfig config;
     private int patchedFileCounter = 0;
 
-    List<? extends SoraldEventHandler> eventHandlers;
+    final List<? extends SoraldEventHandler> eventHandlers;
 
-    public Repair(SoraldConfig config) {
+    public Repair(SoraldConfig config, List<? extends SoraldEventHandler> eventHandlers) {
         this.config = config;
         if (this.config.getGitRepoPath() != null) {
             generator.setGitProjectRootDir(this.config.getGitRepoPath());
         }
         spoonedPath = Paths.get(config.getWorkspace()).resolve(Constants.SPOONED);
         intermediateSpoonedPath = spoonedPath.resolve(Constants.INTERMEDIATE);
-
-        eventHandlers =
-                List.of(
-                        new SoraldEventHandler() {
-                            private long parseStart = -1;
-                            private long parseEnd = -1;
-                            private long repairStart = -1;
-                            private long repairEnd = -1;
-                            private EnumMap<EventType, List<EventMetadata>> allMetadata =
-                                    new EnumMap<>(EventType.class);
-
-                            @Override
-                            public void registerEvent(EventType eventType) {
-                                switch (eventType) {
-                                    case PARSE_START:
-                                        parseStart = System.nanoTime();
-                                        break;
-                                    case PARSE_END:
-                                        parseEnd = System.nanoTime();
-                                        break;
-                                    case REPAIR_START:
-                                        repairStart = System.nanoTime();
-                                        break;
-                                    case REPAIR_END:
-                                        repairEnd = System.nanoTime();
-                                        break;
-                                    default:
-                                        // do nothing
-                                }
-                            }
-
-                            @Override
-                            public void registerEvent(EventType type, EventMetadata metadata) {
-                                registerEvent(type);
-                                List<EventMetadata> eventTypeMetadata =
-                                        allMetadata.getOrDefault(type, new ArrayList<>());
-                                eventTypeMetadata.add(metadata);
-                                allMetadata.putIfAbsent(type, eventTypeMetadata);
-                            }
-
-                            @Override
-                            public void close() {
-                                System.out.println(
-                                        "Time to parse: "
-                                                + (parseEnd - parseStart) / 1_000_000
-                                                + " ms");
-                                System.out.println(
-                                        "Time to repair: "
-                                                + (repairEnd - repairStart) / 1_000_000
-                                                + " ms");
-                                System.out.println(allMetadata);
-                            }
-                        });
+        this.eventHandlers = Collections.unmodifiableList(eventHandlers);
     }
 
     /** Execute a repair according to the config. */
