+++ src/main/java/sorald/SoraldConfig.java
@@ -1,8 +1,10 @@
 package sorald;
 
+import java.io.File;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 /* All config settings of Sorald should be gathered here */
 public class SoraldConfig {
@@ -15,6 +17,7 @@ public class SoraldConfig {
     private String gitRepoPath;
     private int maxFixesPerRule;
     private int maxFilesPerSegment;
+    private File statsOutputFile;
 
     public SoraldConfig() {}
 
@@ -94,5 +97,13 @@ public class SoraldConfig {
 
     public int getMaxFilesPerSegment() {
         return this.maxFilesPerSegment;
+    }
+
+    public void setStatsOutputFile(File statsOutputFile) {
+        this.statsOutputFile = statsOutputFile;
+    }
+
+    public Optional<File> getStatsOutputFile() {
+        return Optional.ofNullable(statsOutputFile);
     }
 }
